% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param1.R
\docType{data}
\name{param_1}
\alias{param_1}
\title{A parameter data set for the \link{view} function with \strong{angular.resolution = 1}}
\description{
A parameter data set for the \link{view} function with \strong{angular.resolution = 1}
}
\keyword{data}
