% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param10.R, R/param_10.R
\docType{data}
\name{param_10}
\alias{param_10}
\title{A parameter data set for the \link{view} function with \strong{angular.resolution = 10}}
\description{
A parameter data set for the \link{view} function with \strong{angular.resolution = 10}

Parameters for the \code{\link{view}} function with an angular resolution of 10 and no gap filling
}
\keyword{data}
