% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param2.R
\docType{data}
\name{param_2}
\alias{param_2}
\title{A parameter for the \link{view} function with \strong{angular.resolution = 2}}
\description{
A parameter for the \link{view} function with \strong{angular.resolution = 2}
}
\keyword{data}
