% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param3.R
\docType{data}
\name{param_3}
\alias{param_3}
\title{A parameter for the \link{view} function with \strong{angular.resolution = 3}}
\description{
A parameter for the \link{view} function with \strong{angular.resolution = 3}
}
\keyword{data}
