% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param4.R
\docType{data}
\name{param_4}
\alias{param_4}
\title{A parameter for the \link{view} function with \strong{angular.resolution = 4}}
\description{
A parameter for the \link{view} function with \strong{angular.resolution = 4}
}
\keyword{data}
