% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param5.R
\docType{data}
\name{param_5}
\alias{param_5}
\title{A parameter for the \link{view} function with \strong{angular.resolution = 5}}
\description{
A parameter for the \link{view} function with \strong{angular.resolution = 5}
}
\keyword{data}
