% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param6.R
\docType{data}
\name{param_6}
\alias{param_6}
\title{A parameter for the \link{view} function with \strong{angular.resolution = 6}}
\description{
A parameter for the \link{view} function with \strong{angular.resolution = 6}
}
\keyword{data}
