% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param7.R
\docType{data}
\name{param_7}
\alias{param_7}
\title{A parameter for the \link{view} function with \strong{angular.resolution = 7}}
\description{
A parameter for the \link{view} function with \strong{angular.resolution = 7}
}
\keyword{data}
