% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param8.R
\docType{data}
\name{param_8}
\alias{param_8}
\title{A parameter for the \link{view} function with \strong{angular.resolution = 8}}
\description{
A parameter for the \link{view} function with \strong{angular.resolution = 8}
}
\keyword{data}
