% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param9.R
\docType{data}
\name{param_9}
\alias{param_9}
\title{A parameter for the \link{view} function with \strong{angular.resolution = 9}}
\description{
A parameter for the \link{view} function with \strong{angular.resolution = 9}
}
\keyword{data}
