% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLSclass.R
\docType{data}
\name{TLSclass}
\alias{TLSclass}
\title{TLS scene of a plantation plot after ground and vegetation segmentation}
\description{
TLS scene of a plantation plot after ground and vegetation segmentation
}
\keyword{data}
